+++ config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  path: {
+    build: {
+      root: 'build',
+      scripts: 'build'
+    },
+    src: {
+      root: 'src',
+      mainScript: 'src/main.js'
+    }
+  }
+}