+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Misc
 - Reorganized structure to prepare for an addition of the app
+- Bootstrapped project with Webpack, Babel, Jest and Standard
 - Added `.editorconfig`
 - Added NPM lockfile.
 - Added changelog.